agent/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via
    #   superior-agent (pyproject.toml)
    #   web3
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.45.2
    # via superior-agent (pyproject.toml)
anyio==4.8.0
    # via
    #   anthropic
    #   httpx
    #   openai
    #   starlette
attrs==25.1.0
    # via aiohttp
beautifulsoup4==4.13.1
    # via bs4
bitarray==3.0.0
    # via eth-account
bs4==0.0.2
    # via superior-agent (pyproject.toml)
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
ckzg==2.0.1
    # via eth-account
click==8.1.8
    # via
    #   duckduckgo-search
    #   uvicorn
cytoolz==1.0.1
    # via eth-utils
decorator==5.1.1
    # via retry
distro==1.9.0
    # via
    #   anthropic
    #   openai
docker==7.1.0
    # via superior-agent (pyproject.toml)
duckduckgo-search==7.3.0
    # via superior-agent (pyproject.toml)
eth-abi==5.2.0
    # via
    #   eth-account
    #   web3
eth-account==0.13.4
    # via web3
eth-hash==0.7.1
    # via
    #   eth-utils
    #   web3
eth-keyfile==0.8.1
    # via eth-account
eth-keys==0.6.1
    # via
    #   eth-account
    #   eth-keyfile
eth-rlp==2.1.0
    # via eth-account
eth-typing==5.1.0
    # via
    #   superior-agent (pyproject.toml)
    #   eth-abi
    #   eth-keys
    #   eth-utils
    #   web3
eth-utils==5.2.0
    # via
    #   eth-abi
    #   eth-account
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   rlp
    #   web3
fastapi==0.115.8
    # via superior-agent (pyproject.toml)
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
hexbytes==1.3.0
    # via
    #   eth-account
    #   eth-rlp
    #   web3
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   superior-agent (pyproject.toml)
    #   anthropic
    #   ollama
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter==0.8.2
    # via
    #   anthropic
    #   openai
loguru==0.7.3
    # via superior-agent (pyproject.toml)
lxml==5.3.0
    # via duckduckgo-search
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   tweepy
ollama==0.4.7
    # via superior-agent (pyproject.toml)
openai==1.60.2
    # via superior-agent (pyproject.toml)
parsimonious==0.10.0
    # via eth-abi
peewee==3.17.8
    # via superior-agent (pyproject.toml)
pip==25.0
    # via superior-agent (pyproject.toml)
polars==1.21.0
    # via superior-agent (pyproject.toml)
primp==0.11.0
    # via duckduckgo-search
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
py==1.11.0
    # via retry
pycryptodome==3.21.0
    # via
    #   eth-hash
    #   eth-keyfile
pydantic==2.10.6
    # via
    #   superior-agent (pyproject.toml)
    #   anthropic
    #   eth-account
    #   fastapi
    #   ollama
    #   openai
    #   pydantic-settings
    #   web3
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.1
    # via superior-agent (pyproject.toml)
pymysql==1.1.1
    # via superior-agent (pyproject.toml)
python-dotenv==1.0.1
    # via
    #   superior-agent (pyproject.toml)
    #   pydantic-settings
pyunormalize==16.0.0
    # via web3
pyyaml==6.0.2
    # via superior-agent (pyproject.toml)
regex==2024.11.6
    # via parsimonious
requests==2.32.3
    # via
    #   superior-agent (pyproject.toml)
    #   docker
    #   requests-oauthlib
    #   tweepy
    #   web3
requests-oauthlib==2.0.0
    # via tweepy
result==0.17.0
    # via superior-agent (pyproject.toml)
retry==0.9.2
    # via superior-agent (pyproject.toml)
rlp==4.0.1
    # via
    #   eth-account
    #   eth-rlp
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   openai
soupsieve==2.6
    # via beautifulsoup4
starlette==0.45.3
    # via fastapi
toolz==1.0.0
    # via cytoolz
tqdm==4.67.1
    # via openai
tweepy==4.15.0
    # via superior-agent (pyproject.toml)
types-requests==2.32.0.20241016
    # via web3
typing-extensions==4.12.2
    # via
    #   anthropic
    #   anyio
    #   beautifulsoup4
    #   eth-typing
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   web3
urllib3==2.3.0
    # via
    #   docker
    #   requests
    #   types-requests
uvicorn==0.34.0
    # via superior-agent (pyproject.toml)
web3==7.8.0
    # via superior-agent (pyproject.toml)
websockets==13.1
    # via web3
yarl==1.18.3
    # via aiohttp
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests

